Updating vulnerable code for centos
On Centos7, pcre8 is available for php. This means the given exploit
won't work as the pcre regex doesn't consider the sender string with the
shell code in it valid.

This means this version of phpmailer isn't vulnerable on this
combo of centos7 + pcre + php, but what I'm building here is a
centos-based image of the given vulnerability, so I'm taking a little
liberty.